requirements.txt
requests==2.25.1
TwitterAPI==2.7.3
PyYAML==5.4.1